pyproject.toml
[tool.poetry]
name = "dexa-sdk"
version = "0.1.3"
description = "Store and manage Data Exchange Agreements (DEXA)"
authors = ["George J Padayatti <george.padayatti@igrant.io>"]
license = "MIT"
packages = [
    { include = "dexa_sdk" }
]

[tool.poetry.dependencies]
python = "^3.8"
asynctest = "^0.13.0"
pytest = "^7.1.2"
green = "^3.4.2"
rich = "^12.5.1"
merkletools = "^1.0.3"
jcs = "^0.2.1"
merklelib = "^1.0"
semver = "^2.13.0"
loguru = "^0.6.0"
pydantic = "^1.9.1"
PyLD = "2.0.1"
rdflib = "^6.2.0"
requests = ">=2.23.0, <2.24.0"
py-multibase = "^1.0.3"
py-multicodec = "^0.2.1"
marshmallow = "3.5.1"
uvloop = "^0.16.0"
MarkupSafe = "2.0.1"
python3-indy = "^1.16.0"
ipykernel = "^6.15.1"
acapy-patched = "0.5.6.dev1"
web3 = "^5.30.0"
py-solc-x = "^1.1.1"
validators = "^0.20.0"
PyJWT = "^2.4.0"
acapy-mydata-did-protocol = { path = "/Users/georgejpadayatti/decentralised-dataexchange/acapy-mydata-did-protocol", develop = true }
dexa-protocol = { path = "/Users/georgejpadayatti/decentralised-dataexchange/dexa-protocol", develop = true }

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
ipykernel = "^6.15.1"
pylint = "^2.14.5"
flake8 = "^5.0.4"
setuptools = "^64.0.3"
mypy = "^0.971"
pdoc3 = "^0.10.0"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"